Major restructure of classes to be consistent with VTK side
