+++ vtki/__init__.py
@@ -11,9 +11,9 @@ from vtki.plotting import *
 from vtki.utilities import *
 from vtki.colors import *
 from vtki.common import Common
-from vtki.polydata import PolyData
-from vtki.grid import UnstructuredGrid
-from vtki.grid import StructuredGrid
+from vtki.pointset import PolyData
+from vtki.pointset import UnstructuredGrid
+from vtki.pointset import StructuredGrid
 from vtki.grid import Grid
 from vtki.grid import RectilinearGrid
 from vtki.grid import UniformGrid
